zdag rework part 2: output linking enforcement, close miner attack vector
output linking creates a graph between all transactions so that users can't pay higher fee to get ahead of placement in blocks, it does CPFP linking through the mempool (senderstatus confirms the graph looks good, still gotta do this).

The miner also was able to do zdag confirm on a tx but mine a different one, for this attack vector we keep track of txids per sender allocation and enforce the clients to only validate ones they have in their mempools. Miner won't be able to include transactions from sender allocations that aren't in peers mempools, and during race conditions peer will sync up from other peers to get the transaction and then revalidate block.


Former-commit-id: e6fe0f811e8f21331da4b029707403234ede92ee

+++ src/init.cpp
@@ -95,7 +95,7 @@ extern AssetBalanceMap mempoolMapAssetBalances;
 extern ArrivalTimesMapImpl arrivalTimesMap; 
 extern CCriticalSection cs_assetallocationmempoolbalance;
 extern CCriticalSection cs_assetallocationarrival;
-extern std::vector<std::pair<uint256, int64_t> >  vecToRemoveFromMempool;
+extern std::vector<std::pair<uint256, uint32_t> >  vecToRemoveFromMempool;
 extern CCriticalSection cs_assetallocationmempoolremovetx;
 static CDSNotificationInterface* pdsNotificationInterface = NULL;
